requirements.txt
requests>=2.32.0
python-dotenv>=1.0.0
tenacity>=8.2.0,<9.0.0
pydantic>=2.9.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<2.3.0
scikit-learn==1.3.2
joblib>=1.3.0
streamlit>=1.38.0
xgboost>=2.0.0
plotly>=5.15.0
pytz>=2024.1
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.14.0
firebase-admin>=6.0.0
torch
